* src/stream_output/sap.c: net_Write() returns an int, not an unsigned.

+++ src/stream_output/sap.c
@@ -540,7 +540,7 @@ static int announce_SAPAnnounceDel( sap_handler_t *p_sap,
 static int announce_SendSAPAnnounce( sap_handler_t *p_sap,
                                      sap_session_t *p_session )
 {
-    unsigned int i_ret;
+    int i_ret;
 
     /* This announce has never been sent yet */
     if( p_session->i_last == 0 )
@@ -558,11 +558,11 @@ static int announce_SendSAPAnnounce( sap_handler_t *p_sap,
         i_ret = net_Write( p_sap, p_session->p_address->i_wfd, NULL,
                            p_session->psz_data,
                            p_session->i_length );
-        if( i_ret != (unsigned int)p_session->i_length )
+        if( i_ret != p_session->i_length )
         {
             msg_Warn( p_sap, "SAP send failed on address %s (%i %i)",
-                   p_session->p_address->psz_address,
-                   i_ret, p_session->i_length );
+                      p_session->p_address->psz_address,
+                      i_ret, p_session->i_length );
         }
         p_session->i_last = p_session->i_next;
         p_session->i_next = p_session->i_last
